ログイン状態の出品者のみ商品を削除

+++ app/controllers/products_controller.rb
@@ -37,8 +37,10 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    if product.destroy
-      redirect_to root_path
+    if current_user.id == product.user_id
+      if product.destroy
+        redirect_to root_path
+      end
     end
   end
